endedupe_integ_test/requirements.txt
attrs==22.2.0
boto3==1.26.106
botocore==1.29.106
coverage==7.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
pytest==7.2.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
